backend/requirements.txt
# ─────────────────────────────────────────
# ShambaFlow Backend Dependencies
# ─────────────────────────────────────────

# Core Framework
Django
djangorestframework

# ── Admin UI — Django Unfold ──────────────
django-unfold
django-import-export

# ── Rich Text Editor — CKEditor 5 ─────────
django-ckeditor-5

# ── API Documentation — Swagger / OpenAPI ─
drf-spectacular               # OpenAPI 3.0 schema generation
drf-spectacular-sidecar

# Database — Neon PostgreSQL
psycopg2-binary
dj-database-url

# Cache — Upstash Redis
redis
django-redis
upstash-redis

# Authentication & Security
djangorestframework-simplejwt
django-cors-headers
cryptography

# Async / WebSockets
channels
channels-redis
daphne

# Email — Brevo (formerly Sendinblue)
sib-api-v3-sdk
requests

# SMS — Infobip
infobip-api-python-client

# File Storage & Media
Pillow
python-magic

# Environment & Config
python-dotenv
django-environ

# Background Tasks
celery
django-celery-beat

# Production Server
gunicorn
gevent

# Utilities
python-slugify
phonenumbers

# Development & Testing
pytest
pytest-django
pytest-cov
factory-boy
django-debug-toolbar